+++ src/main/java/de/alpharogroup/gson/factory/TypeFactory.java
@@ -39,8 +39,7 @@ public final class TypeFactory
 	/**
 	 * Factory method for create a new {@link Type} for convert {@link List} objects
 	 *
-	 * @param <T>
-	 *            the generic type of the list elements
+	 * @param <T>      the generic type of the list elements
 	 * @param listType the list type
 	 * @return the new {@link Type} object
 	 */
@@ -52,13 +51,10 @@ public final class TypeFactory
 	/**
 	 * Factory method for create a new {@link Type} for convert json objects from all derived {@link Collection} objects
 	 *
-	 * @param <T>
-	 *            the generic type of the collection
-	 * @param <E>
-	 *            the generic type of the collection elements
+	 * @param <T>             the generic type of the collection
+	 * @param <E>             the generic type of the collection elements
 	 * @param collectionClass the collection class
-	 * @param elementClass the element class
-	 *
+	 * @param elementClass    the element class
 	 * @return the new {@link Type} object
 	 */
 	public static <T extends Collection, E> Type newCollectionTypeToken(Class<T> collectionClass,
@@ -70,8 +66,7 @@ public final class TypeFactory
 	/**
 	 * Factory method for create a new {@link Type} for convert all derived {@link Collection} objects
 	 *
-	 * @param <T>
-	 *            the generic type of the collection elements
+	 * @param <T>          the generic type of the collection elements
 	 * @param elementClass the element class
 	 * @return the new {@link Type} object
 	 */
@@ -83,16 +78,12 @@ public final class TypeFactory
 	/**
 	 * Factory method for create a new {@link Type} for convert json objects from all derived {@link Map} objects
 	 *
-	 * @param <M>
-	 *            the generic type of the map
-	 * @param <K>
-	 *            the generic type of keys
-	 * @param <V>
-	 *            the generic type of values
-	 * @param mapClass the map class
-	 * @param keyType the key type class
+	 * @param <M>       the generic type of the map
+	 * @param <K>       the generic type of keys
+	 * @param <V>       the generic type of values
+	 * @param mapClass  the map class
+	 * @param keyType   the key type class
 	 * @param valueType the value type class
-	 *
 	 * @return the new {@link Type} object
 	 */
 	public static <M extends Map, K, V> Type newMapTypeToken(Class<M> mapClass, Class<K> keyType,
@@ -104,13 +95,10 @@ public final class TypeFactory
 	/**
 	 * Factory method for create a new {@link Type} for convert json objects from all derived {@link Map} objects
 	 *
-	 * @param <K>
-	 *            the generic type of keys
-	 * @param <V>
-	 *            the generic type of values
-	 * @param keyType the key type class
+	 * @param <K>       the generic type of keys
+	 * @param <V>       the generic type of values
+	 * @param keyType   the key type class
 	 * @param valueType the value type class
-	 *
 	 * @return the new {@link Type} object
 	 */
 	public static <K, V> Type newMapTypeToken(Class<K> keyType, Class<V> valueType)