+++ src/test/java/de/alpharogroup/gson/ObjectToJsonExtensionsTest.java
@@ -32,7 +32,12 @@ import de.alpharogroup.test.objects.enums.Gender;
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
 
 import static org.testng.AssertJUnit.assertEquals;