+++ src/test/java/de/alpharogroup/gson/strategy/FooExclusionStrategy.java
@@ -1,0 +1,18 @@
+package de.alpharogroup.gson.strategy;
+
+import com.google.gson.ExclusionStrategy;
+import com.google.gson.FieldAttributes;
+
+public class FooExclusionStrategy implements ExclusionStrategy
+{
+
+	@Override public boolean shouldSkipClass(Class<?> clazz)
+	{
+		return false;
+	}
+
+	@Override public boolean shouldSkipField(FieldAttributes field)
+	{
+		return field.getAnnotation(FooExclusion.class) != null;
+	}
+}